+++ source/Knyaz.Optimus/Builders/DocumentStylesConfiguration.cs
@@ -1,0 +1,10 @@
+using Knyaz.Optimus.Dom.Css;
+
+namespace Knyaz.Optimus.ResourceProviders
+{
+	public class DocumentStylesConfiguration
+	{
+		/// <summary> The primary style sheet. </summary>
+		public CssStyleSheet UserAgentStyleSheet { get; set; }
+	}
+}